code/requirements.txt

einops==0.6.1
evalutils==0.4.2

imageio==2.31.1

monai==1.1.0

nibabel==5.1.0




SimpleITK

pytorch_lightning

nibabel
